requirements.txt
flask>=2.0.1
flask-cors>=3.0.10
python-dotenv>=0.19.0
solana>=0.29.1
web3>=5.24.0
ipfshttpclient>=0.8.0
numpy>=1.21.2
pandas>=1.3.3
scikit-learn>=0.24.2
pytest>=6.2.5
requests>=2.26.0
python-jose>=3.3.0
